requirements.txt
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
pycocoevalcap==1.2
sentencepiece==0.2.0
open_clip_torch
torch
torchvision
torchaudio
datasets
huggingface-hub
tqdm
wandb
streamlit
psycopg2-binary
dotenv